requirements.txt
asgiref==3.3.1
Django==3.1.5
djangorestframework==3.12.2
flake8==3.8.4
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.5
sqlparse==0.4.1
psycopg2-binary>=2.8
PyJWT
